Time code work
-Made t Float so that slowdown is possible.
-Attempted implementation of using delta time to allow jumping, half works at present
-Fix for Villagers 'Sliding' on death
-Added Within range extension method for Lists and removed static function.
-DebugPnl logic changed so can be activated at start

+++ Assets/Scripts/Boss/Boss Manager.cs
@@ -386,7 +386,7 @@ public abstract class BossManager : MonoBehaviour
 
                                 if (ReqsForNextStage())
                                 {
-                                    TimeEnteredCurrentStage = TimeObjectManager.t;
+                                    TimeEnteredCurrentStage = (int)TimeObjectManager.t;
                                     bossStage = BossStage.Two;
                                     OnStageTwo();
                                 }
@@ -398,7 +398,7 @@ public abstract class BossManager : MonoBehaviour
 
                                 if (ReqsForNextStage())
                                 {
-                                    TimeEnteredCurrentStage = TimeObjectManager.t;
+                                    TimeEnteredCurrentStage = (int)TimeObjectManager.t;
                                     bossStage = BossStage.Three;
                                     OnStageThree();
                                 }
@@ -411,7 +411,7 @@ public abstract class BossManager : MonoBehaviour
 
                                 if (ReqsForNextStage())
                                 {
-                                    TimeEnteredCurrentStage = TimeObjectManager.t;
+                                    TimeEnteredCurrentStage = (int)TimeObjectManager.t;
                                     bossStage = BossStage.Four;
                                     OnStageFour();
                                 }
@@ -424,7 +424,7 @@ public abstract class BossManager : MonoBehaviour
 
                                 if (ReqsForNextStage(false))
                                 {
-                                    TimeEnteredCurrentStage = TimeObjectManager.t;
+                                    TimeEnteredCurrentStage = (int)TimeObjectManager.t;
                                     bossStage = BossStage.Five;
                                     OnStageFive();
                                 }
